+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 1.0.1 -- 2022-06-23
+
+### Added
+
+* `PExchangeRate` - Store exchange rate between two discrete values
+* `pexchangeRatio` - Safely constructs `PExchangeRate`
+* `pexchange` - Perform exchange
+
 ## 1.0.0 -- 2022-05-26
 
 ### Added